Bugfix: install library TARGETS need DESTINATIONs.
See

* CMake 2.8.0 Documentation # Commands:[install](https://cmake.org/cmake/help/v2.8.0/cmake.html#command:install).

You can find descriptions as below:

```
For non-DLL platforms shared libraries are treated
as LIBRARY targets, except that those marked with
the FRAMEWORK property are treated as FRAMEWORK
targets on OS X. For DLL platforms the DLL part
of a shared library is treated as a RUNTIME target
and the corresponding import library is treated
as an ARCHIVE target. All Windows-based systems
including Cygwin are DLL platforms.
```

Because of above reason, in cygwin environment, cmake failes as below:

    ...
    CMake Error at lib/NGT/CMakeLists.txt:20 (install):
      install Library TARGETS given no DESTINATION!

Because, in cygwin environment, the shared library will be installed
into RUNTIME DESTINATION.
So, install library TARGETS also need RUNTIME DESTINATION.