DESTINATION of install library should be `lib` instead of `lib<qual>`.
The 32-bit architecture has no way to execute binary under the `lib64`.
Of course, the 32-bit binary must not be installed under the `lib64`.
On the other hand, the 64-bit binary, it is enough to install it
into the 'lib'.
Maybe the `lib64` should be just a symbolic link to the `lib`.
Maybe the 64-bit architecture should have meaningless to use the `lib64`.